+++ manager2/src/app/tps/tps.component.ts
@@ -3,8 +3,14 @@ import { NgForm } from '@angular/forms';
 import { AuthService } from '../auth/auth.service';
 import { ConfigService } from '../config.service'
 import { TpserviceService } from './tpservice.service';
-import { CalendarEventTimesChangedEvent } from 'angular-calendar';
+import { CalendarEvent, CalendarEventTimesChangedEvent } from 'angular-calendar';
 import { Subject } from 'rxjs';
+
+const eventColors = {
+    created: { primary: '#00dd44' },
+    pending: { primary: '#0066ff' },
+    over: { primary: '#808080' }
+};
 
 @Component({
     selector: 'app-tps',
@@ -24,8 +30,9 @@ export class TpsComponent implements OnInit {
     session_user: any
 
     viewDate: Date
-    events: any
+    events: CalendarEvent[]
     selectedEvent: any
+    refresh: Subject<any> = new Subject();
 
     quantity: number
     fromDate: Date
@@ -35,8 +42,6 @@ export class TpsComponent implements OnInit {
 
     group_or_project: string
     name: string
-
-    refresh: Subject<any> = new Subject();
 
     activeDayIsOpen: boolean = true;
 
@@ -48,17 +53,28 @@ export class TpsComponent implements OnInit {
         private tpService: TpserviceService
     ) { }
 
+    private choseColor(id: string, over: boolean, created: boolean) {
+        let color;
+        if (over) { color = eventColors.over; }
+        else {
+            if (created) { color = eventColors.created; }
+            else { color = eventColors.pending; }
+            const hash = id.split('').reduce((acc, char) => acc + char.charCodeAt(0), 0);
+            const red = (3*hash % 128).toString(16).padStart(2, '0');
+            color.primary = `#${red}${color.primary.slice(3)}`;
+        }
+        return color;
+    }
+
     private listEvents() {
         this.tpService.list().subscribe(
             resp => {
-                let events = [];
-                for(var i = 0; i < resp.length; i++) {
-                    let event = resp[i];
-                    events.push({
-                        'title': event.owner + ', ' + event.quantity + ' students',
-                        'start': new Date(event.from),
-                        'end': new Date(event.to),
-                        'allDay': false,
+                const events: CalendarEvent[] = resp.map(event => {
+                    return {
+                        title: `${event.owner}, ${event.quantity} students`,
+                        start: new Date(event.from),
+                        end: new Date(event.to),
+                        color: this.choseColor(event._id, event.over, event.created),
                         meta: {
                             'id': event._id,
                             'owner': event.owner,
@@ -71,12 +87,12 @@ export class TpsComponent implements OnInit {
                             'group': event.group,
                             'project': event.project
                         }
-                    });
-                }
+                    }
+                });
                 this.events = events;
                 this.refresh.next();
             },
-            err => console.log('failed to log tp reservations')
+            err => console.error('failed to log tp reservations', err)
         );
     }
 
@@ -166,33 +182,39 @@ export class TpsComponent implements OnInit {
         this.msg = '';
         this.errmsg = '';
         this.tpService.cancel(this.selectedEvent.id).subscribe(
-            resp => this.msg = resp['message'],
+            resp => {
+                this.msg = resp['message'];
+                this.selectedEvent.over = true;
+                this.listEvents();
+            },
             err => this.errmsg = err.error.message
         );
-        this.selectedEvent.over = true;
-        this.listEvents();
     }
 
     create_reservation() {
         this.msg = '';
         this.errmsg = '';
         this.tpService.create(this.selectedEvent.id).subscribe(
-            resp => this.msg = resp['message'],
+            resp => {
+                this.msg = resp['message'];
+                this.selectedEvent.created = true;
+                this.listEvents();
+            },
             err => this.errmsg = err.error.message
         );
-        this.selectedEvent.created = true;
-        this.listEvents();
     }
 
     remove_reservation() {
         this.msg = '';
         this.errmsg = '';
         this.tpService.remove(this.selectedEvent.id).subscribe(
-            resp => this.msg = resp['message'],
+            resp => {
+                this.msg = resp['message'];
+                this.selectedEvent.over = true;
+                this.listEvents();
+            },
             err => this.errmsg = err.error.message
         );
-        this.selectedEvent.over = true;
-        this.listEvents();
     }
 
     extend_reservation() {
@@ -208,19 +230,20 @@ export class TpsComponent implements OnInit {
         }
         const extension = { 'to': new Date(this.new_expire).getTime() };
         this.tpService.extend(this.selectedEvent.id, extension).subscribe(
-            resp => this.msg = resp['message'],
+            resp => {
+                this.msg = resp['message'];
+                this.listEvents();
+            },
             err => this.errmsg = err.error.message
         );
-        this.listEvents();
     }
 
     eventClicked(clickedEvent) {
-        let event = clickedEvent.event;
-        this.selectedEvent = event.meta;
-        this.selectedEvent.title = event.title;
-        this.selectedEvent.start = event.start;
-        this.selectedEvent.end = event.end;
-        if(!event.meta.group) {
+        this.selectedEvent = clickedEvent.meta;
+        this.selectedEvent.title = clickedEvent.title;
+        this.selectedEvent.start = clickedEvent.start;
+        this.selectedEvent.end = clickedEvent.end;
+        if (!clickedEvent.meta.group) {
             this.selectedEvent.group = { }
         }
     }